Created hibernate config and made it pass a test.
Hibernate works at this stage and persists the data in a h2 database
created on the fly.

Also made the testPersonCreation() test to compare the toString() method
and made it pass. At this stage the test can be deleted because it is
fully refactored. This will be done in the next commit.

~ Adams Korir